+++ src/environments/environment.ts
@@ -2,4 +2,6 @@ export const environment = {
   authSourceToken: 'YM%Ln#c7Bg94MyAHrs$n4DQnjk$$vErH',
   port: 4000,
   production: false,
+  transportAndLoadService: 'http://localhost:4001/',
+  transportAndLoadToken: 'QBY!qDAuuGHJ8T@b7uga#yvV^PKRYZWz',
 };